separated parallel cg and csr added code for parallel reg
