+++ src/BayesianDataFusion.jl
@@ -7,6 +7,7 @@ include("ROC.jl")
 include("RelationData.jl")
 include("parallel_matrix.jl")
 include("sparsebin_csr.jl")
+include("parallel_csr.jl")
 include("parallel_cg.jl")
 include("macau.jl")
